fix: carry fractional paid minutes through RGDU bridge

+++ app/src/main/java/com/amaury/pointage/v2/EmployerGeneralReductionPayrollBridgeV2.kt
@@ -23,7 +23,7 @@ object EmployerGeneralReductionPayrollBridgeV2 {
         workforceBand: EmployerWorkforceContributionsV2.Band?,
         contractType: ContractTypeV2?,
         contractualWeeklyMinutes: Int?,
-        additionalPaidMinutes: Int?
+        additionalPaidMinutes: Double?
     ): EmployerReductionResolutionV2.Result {
         if (companyId.isBlank()) {
             return EmployerReductionResolutionV2.Result(
